Move GStreamerSurfaceView
